Add fzj code for obj detection
